Implement unboxed sum primitive type
Summary:
This is yet another iteration for the unboxed sums project.

For the reference: the code that implements unboxed sum prim types, `UNPACK`
pragma support for sums, deep demand analysis for sums (i.e. more precise
information than just `HeadStr` for sums + sum CPR), worker/wrapper
transformations for sum types etc. (with many flags for experimenting with
optimizations) are all implemented in
http://github.com/osa1/ghc/tree/rebase-second-try. However, being a research
project with a tight deadline, the code is a mess. In addition, some of the
parts were not done properly or just missing (e.g. we were analysing recursive
sum types in demand analysis because we couldn't implement a working
`isRecursiveTy`) and optimizations didn't work as good as we expected.

Before working more on the optimization pass (demand analysis + worker/wrapper
for sums) we need a stable and efficient compilation of unboxed sums. In this
patch I'm hoping to achieve the "stable" part. Efficient part is a bit more
tricky, as we actually use unboxed tuples under the hood. Implications of using
unboxed tuples under the hood are:

- Pointers and non-pointers can't share same "buckets". (TODO: explain what I
  mean by "bucket")

- Some of the prim types can't share same buckets. For example, `VecRep` values
  can have different sizes, and putting `Int#` and `Float#` to the same bucket
  is causing a panic in the native code generator. (it's generating an illegal
  instruction for moving from a floating point register to a general purpose
  register)

These are causing memory waste when used carelessly, or when the optimizer is
unaware of these stuff and unboxes sum types aggressively. The `Int# | Float#`
case should be fixed by fixing the code generator.

`UNPACK` support is trivial, but I'm leaving that to its own patch as this is
already big enough.

`Parser.y` changes are implemented by Johan Tibell.

TODO: Refer to the Trac ticket.
TODO: Refer to the Wiki page. (the page may need updates)

Reviewers: simonpj, austin, goldfire, bgamari, alanz, simonmar

Subscribers: goldfire, thomie, mpickering

Differential Revision: https://phabricator.haskell.org/D2259