Fix layout of MultiWayIf expressions (#10807)
Summary:
MultiWayIf was entering a layout context based on token `|` that follows
an `if`. This was causing wrong parsing as shown in #10807.

With this patch we record indentation of `if` token in the current
`LayoutContext`, and use recorded indentation when entering a context if
the following token is `|`.

Test Plan:
validate
add new tests

Reviewers: austin, bgamari, simonmar

Subscribers: thomie

Differential Revision: https://phabricator.haskell.org/D2524

GHC Trac Issues: #10807