Remove CCS transformations in SCCfinal
Cost-centres are now initialized correctly in `coreToStg` so no need to
update them in `SCCfinal`.

Also, a simple optimization done in `SCCfinal` (remove ticks if that
makes the closure a con app) is removed because it was already done in
`coreToStg` (and even more aggresively, `SCCfinal` only removes a tick
if it's non-counting, but `coreToStg` removes ticks even if they're
counting, see `stripStgTicksTop` calls in `CoreToStg.hs`).

A StgLint check added to make sure this transformation is done.

`CoredToStg` needs more parameters now, to keep track of whether a let
binding is top-level or not and whether we're building for profiling or
not. This info is used to initialize CCS fields:

  not profiling = noCCS
  profiling and top level = dontCareCCS (see comments in `mkStgRhs'`)
  profiling and not top level = currentCCS