Fix #15038
We introduce a new Id for unused pointer values in unboxed sums that is
not CAFFY. Because the Id is not CAFFY it doesn't make non-CAFFY
definitions CAFFY, fixing #15038.

To make sure anything referenced by the new id will be retained we get a
stable pointer to in on RTS startup.

Alternatively we could implement a static closure in Cmm that calls
`barf()` on entry, but per Simon Marlow's request the previous
exception-throwing semantics is preserved.