Fix #15038
Summary:
We introduce a new Id for unused pointer values in unboxed sums that is
not CAFFY. Because the Id is not CAFFY it doesn't make non-CAFFY
definitions CAFFY, fixing #15038.

To make sure anything referenced by the new id will be retained we get a
stable pointer to in on RTS startup.

Alternatively we could implement a static closure in Cmm that calls
`barf()` on entry, but per Simon Marlow's request the previous
exception-throwing semantics is preserved.

Test Plan:
- Validate
- Figure out a way to add the original reproducer. We spent a lot of time
  trying to make the reproducer smaller while debugging but failed.

Reviewers: bgamari, simonmar, simonpj, hvr, erikd

Subscribers: thomie, carter

GHC Trac Issues: #15038

Differential Revision: https://phabricator.haskell.org/D4652