Fix StgLint bound id check
Summary:
StgLint was incorrectly using isLocalId for bound id check to see whether an id
is imported (in which case we don't expect it to be bound) or local. The problem
with isLocalId is that its semantics changes after Core, as explained in the
note: (last line)

    Note [GlobalId/LocalId]
    ~~~~~~~~~~~~~~~~~~~~~~~
    A GlobalId is
      * always a constant (top-level)
      * imported, or data constructor, or primop, or record selector
      * has a Unique that is globally unique across the whole
        GHC invocation (a single invocation may compile multiple modules)
      * never treated as a candidate by the free-variable finder;
            it's a constant!

    A LocalId is
      * bound within an expression (lambda, case, local let(rec))
      * or defined at top level in the module being compiled
      * always treated as a candidate by the free-variable finder

    After CoreTidy, top-level LocalIds are turned into GlobalIds

We now pass current module as a parameter to StgLint, which uses it to see if an
id should be bound (defined in the current module) or not (imported).

Test Plan:
This validates. I also tried dropping some top-level bindings before running
StgLint. Previously StgLint was accepting the program, thinking that the id
should be bound. Now it rejects the program as expected.

Reviewers: simonpj, bgamari

Subscribers: rwbarton, carter

Differential Revision: https://phabricator.haskell.org/D5370