Refactor GHCi UI to fix #16096
Summary:
Instead of parsing and executing a statement or declaration directly we
now parse them first and then execute in a separate step. This gives us
the flexibility to inspect the parsed declaration before execution.
Using this we now inspect parsed declarations, and if it's a single
declaration of form `x = y` we execute it as `let x = y` instead, fixing
 #16096.

To avoid any users of the modules I left `execStmt` and `runDecls`
unchanged and added `execStmt'` and `runDecls'` which work on parsed
statements/declarations.

TODO:

- There is some code duplications, as indicated by `TODO` comments.
- What a mess ... This stuff needs some serious refactoring, but I'm too
  afraid of breaking downstream... I haven't reviewed this code myself.
- Remove `pprTrace`s.
- I think I broke one test (an error message that was previously printed
  is now missing).

Test Plan: Validate

Reviewers: simonmar, simonpj, RyanGlScott, bgamari

Subscribers: rwbarton, carter

GHC Trac Issues: #16096

Differential Revision: https://phabricator.haskell.org/D5473